TrumpX
TrumpX (TPX) is not just another memecoin  it’s a political statement wrapped in satire and powered by Solana speed.

Fueled by the fire of billionaires and built for the people, TrumpX is here to shake the charts, break the memes, and maybe tweet its way to the moon 🚀

No roadmap. No promises. Just pure decentralized fun, chaos, and maybe... huge gains.

$TPX — Because sometimes crypto just needs a little drama.